refactor(route): test 코드 작성을 위해 server 관련 유틸 분리
Resolves: #264
